Move TopMenu, UserProfile, and Account modals from /admin/users/-components to /shared/components
