+++ application/account-management/WebApp/routes/admin/index.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 import { createFileRoute } from "@tanstack/react-router";
-import { TopMenu } from "@/routes/admin/users/-components/TopMenu";
 import { SideMenu } from "@repo/ui/components/SideMenu";
 import { Trans } from "@lingui/macro";
 import { accountManagementApi } from "@/shared/lib/api/client";
+import { TopMenu } from "@/shared/components/topMenu";
 
 export const Route = createFileRoute("/admin/")({
   component: Home
